open can socket parameter list change
Take the can interface number as a char and converted into a string
that should reflect the can interface name (e.q can0).
As a second parameter take the interface type that can be used for
send or receive CAN frames. The interface set for receive will do
a CAN bus sniffing.
Also did some cleanup in the other functions.